requirements.txt
moondream==0.1.0
pillow==10.4.0
